// app/api/generated/zod.gen.ts
export const zCategory = z.object({
    id: z.coerce.bigint().min(BigInt('-9223372036854775808'), { error: 'Invalid value: Expected int64 to be >= -9223372036854775808' }).max(BigInt('9223372036854775807'), { error: 'Invalid value: Expected int64 to be <= 9223372036854775807' }).optional(),
    name: z.string().optional()
});

export const zTag = z.object({
    id: z.coerce.bigint().min(BigInt('-9223372036854775808'), { error: 'Invalid value: Expected int64 to be >= -9223372036854775808' }).max(BigInt('9223372036854775807'), { error: 'Invalid value: Expected int64 to be <= 9223372036854775807' }).optional(),
    name: z.string().optional()
});

export const zPet = z.object({
    id: z.coerce.bigint().min(BigInt('-9223372036854775808'), { error: 'Invalid value: Expected int64 to be >= -9223372036854775808' }).max(BigInt('9223372036854775807'), { error: 'Invalid value: Expected int64 to be <= 9223372036854775807' }).optional(),
    category: zCategory.optional(),
    name: z.string(),
    photoUrls: z.array(z.string()),
    tags: z.array(zTag).optional(),
    status: z.enum([
        'available',
        'pending',
        'sold'
    ]).optional()
});
// ...
export const zFindPetsByStatusResponse = z.array(zPet);
// ...
export const zGetPetByIdResponse = zPet;