// app/api/generated/types.gen.ts
export type Category = {
    id?: number;
    name?: string;
};

export type Pet = {
    id?: number;
    category?: Category;
    name: string;
    photoUrls: Array<string>;
    tags?: Array<Tag>;
    /**
     * pet status in the store
     */
    status?: 'available' | 'pending' | 'sold';
};

export type Tag = {
    id?: number;
    name?: string;
};